Cleaning up ExpiryTime to be cleaner and allow for easier understanding

+++ SwiftLnd/Extensions/Protobuf+Extensions.swift
@@ -18,7 +18,7 @@ extension Lnrpc_Invoice {
             self.memo = memo
         }
         if let expiry = expiry {
-            self.expiry = expiry.rawValue
+            self.expiry = Int64(expiry.rawValue)
         }
         `private` = true
     }
